refactor(storage): extract connection-owned plugin state queries (#145163)
