+++ src/plugin-state/plugin-state-store.prepared.test.ts
@@ -1,6 +1,10 @@
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
 import { getNodeSqliteKysely } from "../infra/kysely-sync.js";
-import { openOpenClawStateDatabase } from "../state/openclaw-state-db.js";
+import {
+  closeOpenClawStateDatabaseByPath,
+  openOpenClawStateDatabase,
+  runOpenClawStateWriteTransaction,
+} from "../state/openclaw-state-db.js";
 import {
   createOpenClawTestState,
   type OpenClawTestState,
@@ -10,6 +14,7 @@ import {
   createPluginStateSyncKeyedStore,
   resetPluginStateStoreForTests,
 } from "./plugin-state-store.js";
+import { lookupPluginStateEntry, registerPluginStateEntry } from "./plugin-state-store.kernel.js";
 import {
   clearPluginStateStoreForTests,
   seedPluginStateEntriesForTests,
@@ -31,6 +36,43 @@ afterAll(async () => {
 });
 
 describe("plugin state prepared queries", () => {
+  it("uses the supplied connection and keeps registration eviction in its owner's transaction", () => {
+    const scope = { pluginId: "discord", namespace: "owned-kernel" };
+    const defaultStore = createPluginStateSyncKeyedStore<string>(scope.pluginId, {
+      namespace: scope.namespace,
+      maxEntries: 1,
+    });
+    defaultStore.register("original", "default database");
+    const pathname = testState.statePath("kernel-owned.sqlite");
+    const database = openOpenClawStateDatabase({ path: pathname, env: testState.env });
+    const options = { database, env: testState.env };
+    const entry = { ...scope, maxEntries: 1, overflowPolicy: "evict-oldest" as const };
+    runOpenClawStateWriteTransaction(() => {
+      registerPluginStateEntry(database, { ...entry, key: "original", valueJson: '"owned"' }, 10);
+    }, options);
+
+    const aborted = new Error("abort the caller's transaction");
+    expect(() =>
+      runOpenClawStateWriteTransaction(() => {
+        registerPluginStateEntry(
+          database,
+          { ...entry, key: "pending", valueJson: '"pending"' },
+          10,
+        );
+        expect(lookupPluginStateEntry(database, { ...scope, key: "pending" })).toBe("pending");
+        expect(lookupPluginStateEntry(database, { ...scope, key: "original" })).toBeUndefined();
+        throw aborted;
+      }, options),
+    ).toThrow(aborted);
+    expect(lookupPluginStateEntry(database, { ...scope, key: "pending" })).toBeUndefined();
+    expect(lookupPluginStateEntry(database, { ...scope, key: "original" })).toBe("owned");
+    expect(defaultStore.lookup("original")).toBe("default database");
+    closeOpenClawStateDatabaseByPath(pathname);
+    const reopened = openOpenClawStateDatabase({ path: pathname, env: testState.env });
+    expect(lookupPluginStateEntry(reopened, { ...scope, key: "original" })).toBe("owned");
+    expect(lookupPluginStateEntry(reopened, { ...scope, key: "pending" })).toBeUndefined();
+  });
+
   it("compiles exact reads once per connection with fresh scope and expiry bindings", () => {
     const now = Date.now();
     seedPluginStateEntriesForTests([
